requirements.txt
torch
torchvision
scikit-learn
numpy
pillow
opencv-python
fastapi
uvicorn
streamlit